requirements.txt
pandas
numpy
matplotlib
sys
sklearn.mixture